% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertToGrinnID.R
\name{convertToGrinnID}
\alias{convertToGrinnID}
\title{Convert other database IDs to Neo4j internal IDs and grinn IDs}
\usage{
convertToGrinnID(txtInput, nodetype, dbXref)
}
\arguments{
\item{txtInput}{list of keywords containing ids e.g. \code{txtInput} = list('id1', 'id2').
The ids are from a database, see \code{dbXref}.}

\item{nodetype}{string of entity type. It can be one of "metabolite","protein","gene","pathway".}

\item{dbXref}{string of database name. Specify the database name used for the txtInput,
see \code{txtInput}. It can be one of "chebi","kegg","pubchem","inchi","hmdb","smpdb","reactome","uniprot","ensembl","entrezgene".}
}
\value{
data frame of ids. Return empty list if found nothing.
}
\description{
Convert the list of ids to grinn IDs
grinn IDs are recommended to be used in several functions including \code{combineNetwork}, \code{fetchGrinnCorrNetwork}, \code{fetchGrinnModuNetwork},
\code{fetchGrinnDiffCorrNetwork}, \code{fetchCorrGrinnNetwork}, \code{fetchModuGrinnNetwork}, \code{fetchDiffCorrGrinnNetwork}.
}
\examples{
# Convert genes from ENSEMBL ids to grinn ids
txtInput <- list('ENSG00000140459','ENSG00000143811','ENSG00000104524')
result <- convertToGrinnID(txtInput, nodetype="gene", dbXref="ensembl")
# Convert metabolites from chebi ids to Grinn ids
txtInput <- c(371,783)
result <- convertToGrinnID(txtInput, nodetype="metabolite", dbXref="chebi")
# Convert metabolites from kegg ids to Grinn ids, read input from csv or txt file
library(grinn)
data(keggids)
result <- convertToGrinnID(unlist(keggids), nodetype="metabolite", dbXref="kegg")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}

