% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createCyNetwork.R
\name{createCyNetwork}
\alias{createCyNetwork}
\title{\code{createCyNetwork} format resulting network to cytoscapeJS format}
\usage{
createCyNetwork(nodeData, edgeData, nodeColor = "#B0B0B0",
  nodeShape = "ellipse")
}
\description{
Modify the R function from r-cytoscape.js, to return json output for generating Cytoscape.js network
}
\references{
Modify the output generation for Cytoscape.js from \url{https://github.com/cytoscape/r-cytoscape.js/blob/master/cytoscapeJsSimpleNetwork.R}
}
\seealso{
\url{https://github.com/cytoscape/r-cytoscape.js/blob/master/cytoscapeJsSimpleNetwork.R}
}

