% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/curlRequestCypher.R
\name{curlRequestCypher}
\alias{curlRequestCypher}
\title{Execute Cypher query}
\usage{
curlRequestCypher(querystring)
}
\arguments{
\item{querystring}{string of Cypher}
}
\value{
list of quried results. Return \code{NULL} if found nothing.
}
\description{
execute query using Cypher as an input. The function calls curlPerform from \pkg{RCurl} for sending HTTP request to the graph database.
}
\note{
Need to set db location
}
\examples{
# Query metabolites by database id
#querystring <- "UNWIND ['X7','X8'] AS x WITH x MATCH (node:Metabolite) WHERE lower(node.GID) = lower(x) RETURN DISTINCT node"
#result <- curlRequestCypher(querystring)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
The RCurl package \url{http://www.inside-r.org/packages/cran/RCurl}
}
\seealso{
\code{\link{curlRequestUrlToDF}}, \code{\link{curlRequestUrlToList}}, \code{\link{curlPerform}}, \url{http://neo4j.com/docs/milestone/introduction.html}
}

