% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/curlRequestUrlToList.R
\name{curlRequestUrlToList}
\alias{curlRequestUrlToList}
\title{Execute query from URI}
\usage{
curlRequestUrlToList(url)
}
\arguments{
\item{url}{string of URI}
}
\value{
list of quried node or relationships. Return \code{NULL} if found nothing.
}
\description{
execute database query using a URI input. The function calls curlPerform from \pkg{RCurl} for sending HTTP request to the graph database.
}
\examples{
# Get the list of relationships of node 21
#url <- "http://localhost:7474/db/data/node/21/relationships/all"
#result <- curlRequestUrlToList(url)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
The RCurl package \url{http://www.inside-r.org/packages/cran/RCurl}
}
\seealso{
\code{\link{curlRequestCypher}}, \code{\link{curlRequestUrlToDF}}, \code{\link{curlPerform}}, \url{http://neo4j.com/docs/milestone/introduction.html}
}

