% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchNode.R
\name{fetchNode}
\alias{fetchNode}
\title{Query for an entity or list of entities from grinn internal database}
\usage{
fetchNode(txtInput, nodetype, searchField, exactMatch, returnAs, dbXref)
}
\arguments{
\item{txtInput}{list of keywords, can be one of grinn id, name, synonym, database xref, inchi e.g. txtInput = list('name1', 'name2'). 
The type of keywords needs to be provided in the argument \code{searchField}, see \code{searchField}. Default is grinn id e.g. G371.}

\item{nodetype}{string of entity type. It can be one of "metabolite","protein","gene","pathway".}

\item{searchField}{string of keyword property to search for, see \code{txtInput}. It can be one of "grinn","name","synonym","xref","inchi". Default is "grinn".
If \code{searchField = "xref"}, the argument \code{dbXref} is required, see \code{dbXref}.
If \code{searchField = "name" or "synonym"}, the argument \code{exactMatch} is required, see \code{exactMatch}.}

\item{exactMatch}{boolean value. The argument is required if \code{searchField = "name" or "synonym"}, see \code{searchField}. Default is TRUE.}

\item{returnAs}{string of output type. It can be one of "list","json", default is "list".}

\item{dbXref}{string of database name. Specify the database name used for the txtInput when the argument \code{searchField = "xref"}, see \code{txtInput} and \code{searchField}. 
It can be one of "grinn",chebi","kegg","pubchem","hmdb","smpdb","reactome","uniprot","ensembl","entrezgene". Default is "grinn".
If pubchem is used, it has to be pubchem SID (substance ID).}
}
\value{
list of nodes and first neighborhoods. Return empty list if found nothing.
}
\description{
from the list of keywords, query for entities from the grinn internal database. Quried results include entity information and relationship information.
The keywords can be any of these node types: metabolite, protein, gene and pathway.
Relationship information includes information of adjacent nodes, name and additional attributes of relationships.
}
\examples{
# Query metabolites by grinn ids
txtInput <- list('G371','G783')
result <- fetchNode(txtInput, nodetype="metabolite")
# Query genes by KEGG ids
txtInput <- list('hsa:4514','hsa:4537')
result <- fetchNode(txtInput, nodetype="gene", searchField="xref", dbXref="kegg", returnAs="list")
# Query proteins by names
txtInput <- list('14-3-3 protein beta/alpha','ARL14 effector protein-like','6-phosphogluconolactonase')
result <- fetchNode(txtInput, nodetype="protein", searchField="name", returnAs="list")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}

