% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchNodeRelation.R
\name{fetchNodeRelation}
\alias{fetchNodeRelation}
\title{\code{fetchNodeRelation} format get node relationships}
\usage{
fetchNodeRelation(url)
}
\description{
get node relationships as the output for further uses by \code{formatNodeOutput}.
}
\seealso{
\code{formatNodeOutput}
#result <- fetchNodeRelation("http://localhost:7474/db/data/node/53/relationships/in")
#return start-relation-end
}

