% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPtCorrGrinnNetwork.R
\name{fetchPtCorrGrinnNetwork}
\alias{fetchPtCorrGrinnNetwork}
\title{Compute a partial correlation network and expand the network with information from Grinn internal database}
\usage{
fetchPtCorrGrinnNetwork(datX, corrCoef, pval, alpha, epsilon, matrix.completion, returnAs, xTo, filterSource)
}
\arguments{
\item{datX}{data frame containing normalized, quantified omics data e.g. expression data, metabolite intensities. 
Columns correspond to entities e.g. genes, metabolites, and rows to samples e.g. normals, tumors. 
Require 'nodetype' at the first row to indicate the type of entities in each column. See below for details.}

\item{corrCoef}{numerical value to define the minimum value of absolute correlation, from 0 to 1, to include edges in the output.}

\item{pval}{numerical value to define the maximum value of pvalues, to include edges in the output.}

\item{alpha}{a numeric value specifying significance level of each test used in \code{\link{qpAvgNrr}}.}

\item{epsilon}{a numeric value specifying the maximum cutoff value of the non-rejection rate met by the edges that are included in the qp-graph, see \code{\link{qpGraph}}.}

\item{matrix.completion}{a string specifying algorithm to employ in the matrix completion operations used in \code{\link{qpPAC}}}

\item{returnAs}{string of output type. Specify the type of the returned network. 
It can be one of "tab","json","cytoscape", default is "tab". "cytoscape" is the format used in Cytoscape.js}

\item{xTo}{string of node type. It can be one of "metabolite","protein","gene","pathway". See below for details.}

\item{filterSource}{string or list of pathway databases. The argument is required, if \code{xTo = "pathway"}.
The argument value can be any of "SMPDB","KEGG","REACTOME" or combination of them e.g. list("KEGG","REACTOME").}
}
\value{
list of nodes and edges. The list is with the following componens: edges and nodes. Return empty list if found nothing
}
\description{
from input omics data e.g. normalized expression data or metabolomics data, it is a one step function to:

1. Compute a partial correlation network of input omics data using qpgraph functions. 
The correlation coefficients are continuous values between -1 (negative correlation) and 1 (positive correlation), with numbers close to 1 or -1, meaning very closely correlated.

2. Expand the correlation network using information from the Grinn internal database.
The nodes of the correlation network are the keywords input to query the Grinn internal database.
The Grinn internal database contains the networks of the following types that can get expanded to: 
metabolite-protein, metabolite-protein-gene, metabolite-pathway, protein-gene, protein-pathway and gene-pathway, see also \code{\link{fetchGrinnNetwork}}.
}
\details{
datX is matrix in which rows are samples and columns are entities. 

- The correlation network can be expand from datX entites to a specific nodetype, by providing a value to \code{xTo}.

If \code{xTo} is given, the columns of datX are required to use grinn ids for extended queries on the Grinn internal database, see \code{\link{convertToGrinnID}} for id conversion.

If \code{xTo} = NULL , only the correlation network will be returned.
}
\examples{
# Compute a partial correlation network of metabolites and expand to a grinn network of metabolite-protein
dummy <- rbind(nodetype=rep("metabolite"),t(mtcars))
colnames(dummy) <- c('G1.1','G27967','G371','G4.1',paste0('G',sample(400:22000, 28)))
result <- fetchPtCorrGrinnNetwork(datX=dummy, corrCoef=0.7, pval=0.05, returnAs="tab", xTo="protein")
library(igraph)
plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Castelo R. and Roverato A. (2006) A robust procedure for Gaussian graphical model search from microarray data with p larger than n. J. Mach. Learn. Res., 7:2621-2650.

Castelo R. and Roverato A. (2009) Reverse engineering molecular regulatory networks from microarray data with qp-graphs. J Comput Biol, 16(2), pp. 213-27.
}
\seealso{
\pkg{\link{qpgraph}}, \code{\link{qpAvgNrr}}, \code{\link{qpGraph}}, \code{\link{qpPAC}}, \code{\link{fetchGrinnNetwork}}, \url{http://js.cytoscape.org/}
}

