% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchPtCorrNetwork.R
\name{fetchPtCorrNetwork}
\alias{fetchPtCorrNetwork}
\title{\code{fetchPtCorrNetwork} compute partial correlations}
\usage{
fetchPtCorrNetwork(datNorm, alpha = 0.05, threshold = 0.1,
  clusterSize = 8, returnAs = "tab")
}
\description{
The function encapsulates methods from \pkg{qpgraph} including \code{qpAvgNrr} and \code{qpTopPairs}
}
\references{
Castelo R. and Roverato A. (2009). Reverse engineering molecular regulatory networks from microarray data with qp-graphs. J Comput Biol, 16(2), pp. 213-27.
}
\seealso{
\code{\link{qpAvgNrr}}, \code{\link{qpTopPairs}}
library(qpgraph)
library(snow)
library(rlecuyer)
datNorm = read.csv("~/Documents/grinn_sample/lung_miyamoto/metAdj.txt",sep="\t",header=TRUE,row.names=1)
datNorm = datNorm[-1,] #exclude nodetype
dt = sapply(datNorm, function(x) as.numeric(as.character(x)))
nw = fetchPtCorrNetwork(datNorm=dt)
}

