% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchtRelation.R
\name{fetchRelation}
\alias{fetchRelation}
\title{\code{fetchRelation} get path information}
\usage{
fetchRelation(url)
}
\description{
get path information as the output for further uses by \code{formatNetworkOutput}.
}
\seealso{
\code{formatNetworkOutput}
#result <- fetchRelation("http://localhost:7474/db/data/relationship/53")
#return start-relation-end
}

