% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchSubnetwork.R
\name{fetchSubnetwork}
\alias{fetchSubnetwork}
\title{Compute a maximum scoring subnetwork}
\usage{
fetchSubnetwork(statInput, nwInput, fdr, returnAs)
}
\arguments{
\item{statInput}{numeric vector of p-values}

\item{nwInput}{list of nodes and edges}

\item{fdr}{false discovery rate}

\item{returnAs}{string of output type. Specify the type of the returned network. It can be one of "tab","json".}
}
\value{
list of nodes and edges. The list is with the following componens: edges and nodes. Return empty list if error or found nothing.
}
\description{
compute a maximum scoring subnetwork or module by integrating omic data and biological networks.
The function wraps around important features of \pkg{BioNet} including \code{fitBumModel}, \code{scoreNodes}, \code{runFastHeinz}.
}
\details{
The function encapsulates methods from \pkg{BioNet} to compute a functional module or significanly changed subnetworks.
These include: \code{fitBumModel}, \code{scoreNodes} and \code{runFastHeinz}.
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Beisser D., Klau GW., Dandekar T., Müller T. and Dittrich MT. (2010) BioNet: an R-Package for the functional analysis of biological networks. Bioinformatics, 26(8):1129-30

Dittrich MT., Klau GW., Rosenwald A., Dandekar T., Müller T. (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. Bioinformatics, 24(13):i223-31
}
\seealso{
\code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}
kw <- c('G15729','G17561','G16015','G18145','G16708')
grinnNW <- fetchGrinnNetwork(txtInput=kw, from="metabolite", to="protein")
library(grinn)
data(dummyStat)
result <- fetchSubnetwork(statInput = dummyStat, nwInput = grinnNW)
library(igraph)
plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
}

