% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchWGCNAModule.R
\name{fetchWGCNAModule}
\alias{fetchWGCNAModule}
\title{Compute a network module correlated to a phenotypic feature}
\usage{
fetchWGCNAModule(datNorm, datPheno, sfPower, minModuleSize, threshold, returnAs)
}
\arguments{
\item{datNorm}{data frame containing normalized, quantified omics data e.g. expression data, metabolite intensities. 
Columns correspond to entities e.g. genes, metabolites, and rows to samples e.g. normals, tumors. Require 'nodetype' at the first row to indicate the type of entities in each column.}

\item{datPheno}{data frame containing phenotypic data e.g. weight, age, insulin sensitivity. Columns correspond to phenotypes and rows to samples e.g. normals, tumors.}

\item{sfPower}{numerical value of soft-thresholding power for correlation network construction. It is automatically estimated using \code{pickSoftThreshold}, or it can be defined by users.}

\item{minModuleSize}{numerical value of minimum module size for module detection.}

\item{threshold}{numerical value to define the minimum value of similarity threshold, from 0 to 1, to include edges in the output.}

\item{returnAs}{string of output type. Specify the type of the returned network. 
It can be one of "tab","json","cytoscape". "cytoscape" is the format used in Cytoscape.js}
}
\value{
list of nodes and edges. The list is with the following componens: edges and nodes. Return empty list if found nothing
}
\description{
identify correlation between the input omics data e.g. normalized gene expression data, and phenotypic data e.g. weight.
The function wraps around important aspects of \pkg{WGCNA} including \code{blockwiseModules}, \code{cor}, \code{corPvalueStudent}, \code{labeledHeatmap}.
These aspects automatically perform correlation network construction, module detection, and display module-phenotype correlations.
A module or the combination of modules can be selected from the heatmap of module-phenotype correlations for including in the network output, see more details below.
}
\details{
The function encapsulates several methods from \pkg{WGCNA} so that module-phenoty correlation analysis can be fasten. These methods include:

- \code{pickSoftThreshold} estimates soft-thresholding powers from scale free topology to build the correlation network.

- \code{blockwiseModules} automatically calculates a correlation network and detects modules.
Modules are the areas of the network where nodes are densely connected based on their topological overlap measure, see \pkg{WGCNA} for more details. 
Each module is labeled by color. By using the color, a module or the combination of modules can be selected ("enter color to the terminal"), for including in the network output.

- Module-phenotype correlations and significances are calculated using WGCNA functions \code{cor} and \code{corPvalueStudent}.
The correlation coefficients are continuous values between -1 (negative correlation) and 1 (positive correlation), with numbers close to 1 or -1, meaning very closely correlated.

- \code{labeledHeatmap} plots a heatmap of module-phenotype correlations. A row is a module and a column is a phynotype. 
Each cell presents the corresponding correlation and the pvalue inside parenthesis. Each cell is colored by correlation, red means positive and blue means negative correlation.

- \code{exportNetworkToCytoscape} exports a network for using in Cytoscape (\url{http://cytoscape.org/}).
The selected module is exported as the network output in which an edge will be included if it similarity threshold above the cutoff, see \code{threshold}.
}
\examples{
# Compute a correlation of metabolite module to phenotypic data
library(grinn)
data(dummy)
data(dummyPheno)
result <- fetchWGCNAModule(datNorm=dummy, datPheno=dummyPheno, minModuleSize=5, threshold=0.2)
# enter module color(s) seperate by space:yellow brown purple
#library(igraph)
#plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Langfelder P. and Horvath S. (2008) WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics, 9:559

Dudoit S., Yang YH., Callow MJ. and Speed TP. (2002) Statistical methods for identifying differentially expressed genes in replicated cDNA microarray experiments, STATISTICA SINICA, 12:111

Langfelder P. and Horvath S. Tutorials for the WGCNA package \url{http://labs.genetics.ucla.edu/horvath/CoexpressionNetwork/Rpackages/WGCNA/Tutorials/index.html}
}
\seealso{
\code{\link{cor}}, \code{\link{corPvalueStudent}}, \code{\link{pickSoftThreshold}}, \code{\link{blockwiseModules}}, \code{\link{labeledHeatmap}}, \code{\link{exportNetworkToCytoscape}}, \url{http://js.cytoscape.org/}
}

