% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNetworkOutput.R
\name{formatNetworkOutput}
\alias{formatNetworkOutput}
\title{\code{formatNetworkOutput} format resulting network}
\usage{
formatNetworkOutput(network, returnAs)
}
\description{
Format network output for further uses by \code{fetchGrinnNetwork}. Retrieve also node relationships using \code{fetchNodeRelation}.
}
\seealso{
\code{fetchGrinnNetwork}, \code{fetchNodeRelation}
#querystring="UNWIND ['G371','G783'] AS x WITH x MATCH ptw = (from:Metabolite)<-[rel:MAP_TO]-(to:Pathway) 
#WHERE lower(from.GID) = lower(x) RETURN DISTINCT ptw"
#network = curlRequestCypher(querystring)
#formatNetworkOutput(network,returnAs="tab")
}

