% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatWGCNAModule.R
\name{formatWGCNAModule}
\alias{formatWGCNAModule}
\title{\code{formatWGCNAModule} format resulting WGCNA module}
\usage{
formatWGCNAModule(datX, softPower, mdColors, threshold, nodetype, returnAs)
}
\description{
Modify the R codes from WGCNA tutorials, to return module output for further uses by 
\code{fetchGrinnModuNetwork}, \code{fetchModuGrinnNetwork}, \code{fetchWGCNAModule}.
}
\references{
Langfelder P. and Horvath S. (2008) WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics, 9:559 
Dudoit S., Yang YH., Callow MJ. and Speed TP. (2002) Statistical methods for identifying differentially expressed genes in replicated cDNA microarray experiments, STATISTICA SINICA, 12:111
Langfelder P. and Horvath S. Tutorials for the WGCNA package \url{http://labs.genetics.ucla.edu/horvath/CoexpressionNetwork/Rpackages/WGCNA/Tutorials/index.html}
}
\seealso{
\code{\link{exportNetworkToCytoscape}}, \code{\link{TOMsimilarityFromExpr}}, \code{fetchGrinnModuNetwork}, \code{fetchModuGrinnNetwork}, \code{fetchWGCNAModule}
}

