% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGrinnDb.R
\name{getGrinnDb}
\alias{getGrinnDb}
\title{Get grinn DB location}
\usage{
getGrinnDb()
}
\value{
string grinn DB location.
}
\description{
get grinn DB location of the current working envornment.
By default grinn DB lacation is http://grinn.genomecenter.ucdavis.edu:7474/db/data/cypher
which contains only human database.
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}

