% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/queryNode.R
\name{queryNode}
\alias{queryNode}
\title{\code{queryNode} query for nodes}
\usage{
queryNode(querystring)
}
\arguments{
\item{querystring}{string of Cypher}
}
\value{
list of data.frame containing node information and repationship information. Return empty list if found nothing.
}
\description{
query for nodes from a Cypher input. Quried results include node information and relationship information.
Relationship information includes ID and name of adjacent nodes, name and additional attributes of relationships
}
\note{
use in grinnWeb only
}
\examples{
# Query metabolites by name
#querystring = "UNWIND ['coa','co2','Malonyl-CoA'] AS x WITH x MATCH (n:Metabolite) WHERE lower(n.name) = lower(x) RETURN DISTINCT n"
#result = queryNode(querystring)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
The RCurl package \url{http://www.inside-r.org/packages/cran/RCurl}
}
\seealso{
\code{\link{curlRequestCypher}}, \code{\link{curlPerform}}
}

