% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setGrinnDb.R
\name{setGrinnDb}
\alias{setGrinnDb}
\title{Set grinn DB location}
\usage{
setGrinnDb()
}
\arguments{
\item{url}{grinn DB location}
}
\description{
set grinn DB location for the currently working environment.
}
\examples{
setGrinnDb("http://localhost:7474/db/data/cypher")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}

