% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setGrinnDb.R
\name{setGrinnDb}
\alias{setGrinnDb}
\title{Set Grinn database location}
\usage{
setGrinnDb()
}
\arguments{
\item{url}{of Grinn database location}
}
\description{
set Grinn database location for the currently working environment.
}
\examples{
# Set Grinn database location to local server
setGrinnDb("http://localhost:7474/db/data/")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}

